Rename lockUnlock.js to lock-unlock.js
