+++ packages/editor/src/lock-unlock.js
